requirements.txt
# Function dependencies, for example:
# package>=version
pandas>=1.3.5
pandas_gbq>=0.17.9
google-cloud-storage>=2.5.0
